Move calculateFixedPoint to resize module
